Mainly unit test changes PscDiscrepancy Added toString.
PscSubmission
Added toString.
Fixed equals to use deepEquals on discrepancies.

PscDiscrepancyReportService
When update causes an attempt to submit a report to CHIPS because the
status is COMPLETE, send that version of the report back in the response
to the UPDATE, not the version that is saved after attempting to send to
CHIPS.
Moved code that handles onReportCompleted into sendReport method and
 renamed it to onReportCompleted.
In onReportCompleted, catch-and-swallow MongoEx, with logging.
In onReportCompleted, use try-with-resources to close http client.
Use ReportStatus enum rather than strings.